# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a specialized server for managing trading strategies and margins.",
        "title": "Trades Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/price-action/fvg": {
            "post": {
                "tags": [
                    "PriceAction (Admin)"
                ],
                "summary": "Save FVG",
                "parameters": [
                    {
                        "description": "FVG Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ObRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "PriceAction (Admin)"
                ],
                "summary": "Delete FVG",
                "parameters": [
                    {
                        "description": "Delete Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ObRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "PriceAction (Admin)"
                ],
                "summary": "Update FVG",
                "parameters": [
                    {
                        "description": "Update Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ObRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/price-action/fvg/check": {
            "post": {
                "description": "Triggers a fresh scan of all stocks against saved Fair Value Gaps (FVG) to identify active mitigations. This bypasses the cache and updates it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PriceAction"
                ],
                "summary": "Force Refresh FVG Mitigations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/price-action/fvg/cleanup": {
            "post": {
                "description": "Triggers a maintenance task that fetches NSE history for all stored symbols and deletes FVGs that have been breached or filled by subsequent price action.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PriceAction"
                ],
                "summary": "Clean up filled FVGs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/price-action/fvg/old/{stopDate}": {
            "post": {
                "description": "Upload a CSV to find and save FVGs from a specific stop date backwards.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PriceAction"
                ],
                "summary": "Process Historical Fair Value Gaps (FVG)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stop Date (YYYY-MM-DD)",
                        "name": "stopDate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/price-action/ob": {
            "post": {
                "tags": [
                    "PriceAction (Admin)"
                ],
                "summary": "Save OB",
                "parameters": [
                    {
                        "description": "OB Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ObRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "PriceAction (Admin)"
                ],
                "summary": "Delete OB",
                "parameters": [
                    {
                        "description": "Delete Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ObRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "PriceAction (Admin)"
                ],
                "summary": "Update OB",
                "parameters": [
                    {
                        "description": "Update Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ObRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/price-action/ob/check": {
            "post": {
                "description": "Triggers a fresh scan of all stocks against saved Order Blocks to find active mitigations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PriceAction"
                ],
                "summary": "Force Refresh OB Mitigations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
        "/price-action/ob/old/{stopDate}": {
            "post": {
                "description": "Upload a CSV to find and save Order Blocks from a specific stop date backwards.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PriceAction"
                ],
                "summary": "Process Historical Order Blocks",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stop Date (YYYY-MM-DD)",
                        "name": "stopDate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/price-action/{symbol}": {
            "get": {
                "tags": [
                    "PriceAction"
                ],
                "summary": "Get PA by Symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Symbol",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ObRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Update successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}